+++ src/restfw/config/sub_resources.py
@@ -2,19 +2,24 @@
 :Authors: cykooz
 :Date: 12.01.2021
 """
+from pyramid.config import Configurator
 from zope.interface.verify import verifyObject
 
 from .common import derive_fabric
 from .. import interfaces
 
 
-def add_sub_resource_fabric(config, fabric, name, parent=interfaces.IResource, **predicates):
+def add_sub_resource_fabric(
+        config: Configurator,
+        fabric,
+        name: str,
+        parent=interfaces.IResource,
+        add_link_into_embedded=False,
+        **predicates,
+):
     """A configurator command for register sub-resource fabric.
-    :type config: pyramid.config.Configurator
-    :type fabric: interfaces.ISubResourceFabric
-    :type name: str
-    :type parent: type or zope.interface.Interface
-    :type predicates: dict
+    :param fabric: Object that provides interfaces.ISubResourceFabric
+    :param parent: type or zope.interface.Interface
 
     Using the default ``parent`` value, ``IResource`` will cause the sub resource
     fabric to be registered for all resource types.
@@ -42,6 +47,7 @@ def add_sub_resource_fabric(config, fabric, name, parent=interfaces.IResource, *
     )
     intr['fabric'] = fabric
     intr['parent'] = parent
+    intr['add_link_into_embedded'] = add_link_into_embedded
 
     def register():
         pred_list = config.get_predlist('sub_resource_fabric')
@@ -55,10 +61,27 @@ def add_sub_resource_fabric(config, fabric, name, parent=interfaces.IResource, *
             'derived_fabric': derived_fabric,
         })
 
-        config.registry.registerAdapter(derived_fabric, required=parent, provided=interfaces.IResource, name=name)
+        config.registry.registerAdapter(
+            derived_fabric,
+            required=parent,
+            provided=interfaces.IResource,
+            name=name,
+        )
+        if add_link_into_embedded:
+            add_into_embedded_fabric = derive_fabric(_add_into_embedded_fabric, preds)
+            config.registry.registerAdapter(
+                add_into_embedded_fabric,
+                required=parent,
+                provided=interfaces.IAddSubresourceLinkIntoEmbedded,
+                name=name,
+            )
 
     config.action(None, register, introspectables=(intr,))
     return fabric
+
+
+def _add_into_embedded_fabric(resource):
+    return True
 
 
 def add_sub_resource_fabric_predicate(